Add graffiti limit
Even though we set maximum length of graffiti in setup wizard, that value can be edited through config tab after package is installed. If value is longer than 32 characters, validator will not start and that is a pretty significant issue. This ensures that graffiti message is cropped and that validator starts